Version 1 (#24)
* Added the class to standardize JOSN responses from the API. Note there are no tests for this class implementation (#10)

* Created ruby script to parse csv filees and generate types for database models (#12)

Great work, you are so good Will this will be used all the time +A

* Ruby updates (#13)

* Changes to the nomenclature to be consistant with sequelize

* Fixed the typo in the require line added by previous commit

* Edited the csv parser to more accurately detect float values and to take command line arguments (#15)

* Expirement (#16)

* Added the inital changes to include the experiment

Fixed the style settings to allow async function notation as well as express naming conventions.
Added the layout pug that was not commited with the error pug for handling 500s
created the migrations and models for experiments
routes for create delete and get have been made but need some changes before final

* Response Updates and expirements polish

Updated the response to keep arrays in array format instead of converting into objects
Changed the / path on expirement route to query all expirements
Cleaned the file and verified comments

This should be a good version 1 of the expirement routes, tests still need to be added

* Target (#18)

* Commiting the changes made from the package-lock.json when I npm installed our application

* created target model and migration using the csv parser

* Fixed some formatting issues caused by using an older version of the cli

* Router boilerplate for target is finished. Postman tests have confirmed each route.

* Created target_id model and migration

* Ruby Script Update

Txt file now generated in the folder of the csv file. Added a new script to the npm scripts. parsing can be done through npm run parse -- <filename> (#17)

* Header update to allow cross site (#19)

Changed the headers to allow cors and changed default port to allow running front and back
ends on the same system

* Target (#20)

* Commiting the changes made from the package-lock.json when I npm installed our application

* created target model and migration using the csv parser

* Fixed some formatting issues caused by using an older version of the cli

* Router boilerplate for target is finished. Postman tests have confirmed each route.

* Created target_id model and migration

* added primary key to model to keep nodemon from crashing

* Ruby updates to include multiple output types. (#22)

UML JSON and SQL are the current output format types
they now output to files with endings matching the output type

* Added Citation (#21)

Citation model and citation id model have been added in.
Routes for citation are in but citation id will remain just for relational backend use

* User (#23)

* Added passport to the project as well as bcrypt and passport-jwt. Added basic passport configuration file to setup our passport strategy.

* Added the user table, implemented password hashing using bcrypt when you save a user, and implemented the standard basic routes as well as a login route that returns a jwt upon successful login

* Added router for user to the server. began implementing authentication but didn't quite get it finished. We will probbably need to rethink our routing scheme and move login and create user to another router without authentication and put authentication on everything else